Davyd/history improvment

+++ Backend/WriteAPI/BI/Services/History/HistoryService.cs
@@ -13,7 +13,9 @@ namespace BI.Services.History
     public class HistoryService
     {
         private HistoryCacheService historyCacheService;
+
         IServiceScopeFactory serviceScopeFactory;
+
         public HistoryService(HistoryCacheService historyCacheServicу, IServiceScopeFactory serviceScopeFactory)
         {
             this.historyCacheService = historyCacheServicу;
@@ -25,9 +27,21 @@ namespace BI.Services.History
             try
             {
                 var histories = historyCacheService.GetCacheHistoriesForSnapshotingByTime(10);
-                // TODO BUG
-                // MakeCopy(histories).Wait();
-                historyCacheService.RemoveFromList(histories);
+                if (histories.Any())
+                {
+                    try
+                    {
+                        MakeCopy(histories).Wait();
+                    }
+                    catch(Exception e)
+                    {
+                        Console.WriteLine(e);
+                    }
+                    finally
+                    {
+                        historyCacheService.RemoveFromList(histories);
+                    }
+                }
             }catch(Exception e)
             {
                 Console.WriteLine(e);
@@ -38,27 +52,11 @@ namespace BI.Services.History
         {
             using var scope = serviceScopeFactory.CreateScope();
             var _mediator = scope.ServiceProvider.GetService<IMediator>();
-            var noteHistoryRepository = scope.ServiceProvider.GetService<NoteHistoryRepository>();
-            var userNoteHistoryManyToManyRepository = scope.ServiceProvider.GetService<UserNoteHistoryManyToManyRepository>();
+            var noteHistoryRepository = scope.ServiceProvider.GetService<NoteSnapshotRepository>();
             foreach (var history in histories)
             {
-                var command = new CopyNoteCommand().GetIsHistory(history.AuthorNoteEmail, new List<Guid> { history.NoteId });
+                var command = new MakeNoteHistoryCommand(history.NoteId, history.UsersThatEditIds);
                 var results = await _mediator.Send(command);
-                var result = results.First();
-                var noteHistory = new NoteHistory()
-                {
-                    NoteId = history.NoteId,
-                    SnapshotTime = DateTimeOffset.Now,
-                    // TODO
-                    // NoteVersionId = result.Id 
-                };
-                var dbNoteHistory = await noteHistoryRepository.Add(noteHistory);
-                var userHistories = history.UsersThatEditIds.Select(userId => new UserNoteHistoryManyToMany()
-                {
-                    NoteHistoryId = dbNoteHistory.Entity.Id,
-                    UserId = userId
-                });
-                await userNoteHistoryManyToManyRepository.AddRange(userHistories);
             }
         }
